Refactoring components and folders
