--- src/app/header/page.tsx
@@ -1,14 +1,0 @@
-import React from 'react'
-import Navbar from './_modules/Navbar'
-
-const page = () => {
-  return (
-    <>
-        <div>
-            <Navbar />
-        </div>
-    </>
-  )
-}
-
-export default page
