+++ src/app/page.tsx
@@ -1,8 +1,11 @@
+import Navbar from '@/components/header/Navbar';
 import { Button } from '@/components/ui/button';
 
 export default function Home() {
-	return <main>
-
+	return (
+		<main>
 		
-	</main>;
+		
+		</main>
+	)
 }
